// packages/battle/src/v2/schemas/skill.schema.d.ts
import { type Static } from '@sinclair/typebox';
export declare const MultihitSchema: import("@sinclair/typebox").TUnion<[import("@sinclair/typebox").TTuple<[import("@sinclair/typebox").TNumber, import("@sinclair/typebox").TNumber]>, import("@sinclair/typebox").TNumber]>;
export declare const BaseSkillSchema: import("@sinclair/typebox").TObject<{
    type: import("@sinclair/typebox").TLiteral<"baseSkill">;
    id: import("@sinclair/typebox").TString;
    sfxRef: import("@sinclair/typebox").TOptional<import("@sinclair/typebox").TString>;
    category: import("@sinclair/typebox").TUnion<import("@sinclair/typebox").TLiteral<string>[]>;
    element: import("@sinclair/typebox").TUnion<import("@sinclair/typebox").TLiteral<string>[]>;
    power: import("@sinclair/typebox").TNumber;
    accuracy: import("@sinclair/typebox").TNumber;
    rage: import("@sinclair/typebox").TNumber;
    priority: import("@sinclair/typebox").TNumber;
    target: import("@sinclair/typebox").TUnion<import("@sinclair/typebox").TLiteral<string>[]>;
    multihit: import("@sinclair/typebox").TUnion<[import("@sinclair/typebox").TTuple<[import("@sinclair/typebox").TNumber, import("@sinclair/typebox").TNumber]>, import("@sinclair/typebox").TNumber]>;
    sureHit: import("@sinclair/typebox").TBoolean;
    sureCrit: import("@sinclair/typebox").TBoolean;
    ignoreShield: import("@sinclair/typebox").TBoolean;
    ignoreOpponentStageStrategy: import("@sinclair/typebox").TUnion<import("@sinclair/typebox").TLiteral<string>[]>;
    tags: import("@sinclair/typebox").TArray<import("@sinclair/typebox").TString>;
    effectIds: import("@sinclair/typebox").TArray<import("@sinclair/typebox").TString>;
}>;
export declare const SkillSchema: import("@sinclair/typebox").TObject<{
    type: import("@sinclair/typebox").TLiteral<"skill">;
    id: import("@sinclair/typebox").TString;
    baseSkillId: import("@sinclair/typebox").TString;
    ownerId: import("@sinclair/typebox").TOptional<import("@sinclair/typebox").TString>;
    category: import("@sinclair/typebox").TUnion<import("@sinclair/typebox").TLiteral<string>[]>;
    element: import("@sinclair/typebox").TUnion<import("@sinclair/typebox").TLiteral<string>[]>;
    target: import("@sinclair/typebox").TUnion<import("@sinclair/typebox").TLiteral<string>[]>;
    multihit: import("@sinclair/typebox").TUnion<[import("@sinclair/typebox").TTuple<[import("@sinclair/typebox").TNumber, import("@sinclair/typebox").TNumber]>, import("@sinclair/typebox").TNumber]>;
    sureHit: import("@sinclair/typebox").TBoolean;
    sureCrit: import("@sinclair/typebox").TBoolean;
    ignoreShield: import("@sinclair/typebox").TBoolean;
    tags: import("@sinclair/typebox").TArray<import("@sinclair/typebox").TString>;
    effectIds: import("@sinclair/typebox").TArray<import("@sinclair/typebox").TString>;
    appeared: import("@sinclair/typebox").TBoolean;
}>;
export type BaseSkillData = Static<typeof BaseSkillSchema>;
export type SkillData = Static<typeof SkillSchema>;
//# sourceMappingURL=skill.schema.d.ts.map